// src/app/model/cita.ts
export class Cita {
    id: number = 0;
    fecha: Date = new Date();
    turno: string = '';
    precio: number = 0.0;
    dniPaciente: string = '';
    idMedico: number = 0;
    pacienteId: number = 0;
    medicoNombre: string = '';
    medicoEspecialidad: string = '';
    medicoHospital: string = '';
    nombrePaciente: string = '';
  }
  